Original code I was given to work with: "You are provided with an abstract class, which also happens to implement an interface, which is also given to you.
A SequenceSearch object, once constructed, contains a content String as well as a start-tag and an end-tag.

The constructor ensures that the content is not null and while the start and end tag are allowed to be the same, they cannot contain each other.
The class' abstract methods are suppose to find all tagged sequences in the content and identify the longest tagged sequence.

The content is considered syntactically correct and simple. You don't need to worry about nested tags and every start-tag is followed by an end-tag."